backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pypdf
numpy
faiss-cpu
sentence-transformers
requests
python-multipart
